// SFU-HUB-CMPT276-main/apps/backend/src/@types/openapi.d.ts
export interface components {
    schemas: {
        EventPost: {
            id?: string;
            allDay?: boolean;
            /** Format: date-time */
            start: string;
            /** Format: date-time */
            end: string;
            title: string;
            url?: string;
            color?: string;
            tags: string[];
        };
        Event: {
            id: string;
            allDay?: boolean;
            /** Format: date-time */
            start: string;
            /** Format: date-time */
            end: string;
            title: string;
            url?: string;
            color?: string;
            tags: string[];
        };
        Resource: {
            /** @description Unique identifier */
            id: string;
            /** @description Title of the resource */
            title: string;
            /** @description Description of the resource */
            description?: string;
            /** @description Link to the specific resource */
            link: string;
            /** @description Category of the resource */
            category: string;
        };
        Restaurants: {
            /** @description Unique identifier */
            id: string;
            /** @description Name of the dining option */
            name: string;
            /** @description URL to an image of the restaurant */
            image: string;
            /** @description Address of the dining location */
            address: string;
            /** @description Location of the restaurant relative to sfu */
            category: string;
            /** @description Contact information */
            contact: string;
            /** @description Description of the dining option */
            description: string;
            /** @description Website URL */
            website: string;
            /** @description Operating hours */
            hours: string;
            /** @description URL to the menu image or page */
            menu: string;
        };
        TransitStop: {
            /** @description Unique identifier for the transit stop. */
            stopNumber: string;
            /** @description Name of the transit stop. */
            stopName: string;
            /** @description List of scheduled stops at this transit stop. */
            stops: components["schemas"]["TransitRoute"][];
        };
        TransitRoute: {
            /** @description Destination of the transit route. */
            destination: string;
            /** @description Name of the transit route. */
            routeName: string;
            /** @description Code of the transit route. */
            routeCode: string;
            /** @description Scheduled arrival time as a Unix timestamp. */
            time: number;
        };
        Error: {
            /**
             * Format: int32
             * @description The HTTP status code
             */
            code: number;
            /** @description The error */
            message: string;
        };
        /** @description An HTTP Problem Details object, as defined in IETF RFC 7807 (https://tools.ietf.org/html/rfc7807). */
        Problem: {
            /** @description The details of the problem */
            detail?: string;
            /** @description The HTTP status code */
            status: number;
            /** @description The type of error (text) */
            title: string;
            /** @description The type of error (URL) */
            type?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}